Move common uart code into arch_driver Component
