+++ Sming/Arch/Esp32/Components/driver/uart.cpp
@@ -152,36 +152,6 @@ smg_uart_t* get_physical(smg_uart_t* uart)
 	return uart;
 }
 
-bool realloc_buffer(SerialBuffer*& buffer, size_t new_size)
-{
-	if(buffer != nullptr) {
-		if(new_size == 0) {
-			smg_uart_disable_interrupts();
-			delete buffer;
-			buffer = nullptr;
-			smg_uart_restore_interrupts();
-			return true;
-		}
-
-		return buffer->resize(new_size) == new_size;
-	}
-
-	if(new_size == 0) {
-		return true;
-	}
-
-	// Avoid allocating in SPIRAM
-	auto mem = heap_caps_malloc(sizeof(SerialBuffer), MALLOC_CAP_DEFAULT | MALLOC_CAP_INTERNAL);
-	auto new_buf = new(mem) SerialBuffer;
-	if(new_buf != nullptr && new_buf->resize(new_size) == new_size) {
-		buffer = new_buf;
-		return true;
-	}
-
-	delete new_buf;
-	return false;
-}
-
 /** @brief UART interrupt service routine
  *  @note both UARTS share the same ISR, although UART1 only supports transmit
  */
@@ -314,51 +284,6 @@ void smg_uart_set_callback(smg_uart_t* uart, smg_uart_callback_t callback, void*
 		uart->param = param;
 		uart->callback = callback;
 	}
-}
-
-size_t smg_uart_resize_rx_buffer(smg_uart_t* uart, size_t new_size)
-{
-	if(smg_uart_rx_enabled(uart)) {
-		realloc_buffer(uart->rx_buffer, new_size);
-	}
-	return smg_uart_rx_buffer_size(uart);
-}
-
-size_t smg_uart_rx_buffer_size(smg_uart_t* uart)
-{
-	return uart != nullptr && uart->rx_buffer != nullptr ? uart->rx_buffer->getSize() : 0;
-}
-
-size_t smg_uart_resize_tx_buffer(smg_uart_t* uart, size_t new_size)
-{
-	if(smg_uart_tx_enabled(uart)) {
-		realloc_buffer(uart->tx_buffer, new_size);
-	}
-	return smg_uart_tx_buffer_size(uart);
-}
-
-size_t smg_uart_tx_buffer_size(smg_uart_t* uart)
-{
-	return uart != nullptr && uart->tx_buffer != nullptr ? uart->tx_buffer->getSize() : 0;
-}
-
-int smg_uart_peek_char(smg_uart_t* uart)
-{
-	return uart != nullptr && uart->rx_buffer ? uart->rx_buffer->peekChar() : -1;
-}
-
-int smg_uart_rx_find(smg_uart_t* uart, char c)
-{
-	if(uart == nullptr || uart->rx_buffer == nullptr) {
-		return -1;
-	}
-
-	return uart->rx_buffer->find(c);
-}
-
-int smg_uart_peek_last_char(smg_uart_t* uart)
-{
-	return uart != nullptr && uart->rx_buffer != nullptr ? uart->rx_buffer->peekLastChar() : -1;
 }
 
 size_t smg_uart_read(smg_uart_t* uart, void* buffer, size_t size)
@@ -687,7 +612,7 @@ smg_uart_t* smg_uart_init_ex(const smg_uart_config_t& cfg)
 	auto txBufferSize = cfg.tx_size;
 
 	if(smg_uart_rx_enabled(uart)) {
-		if(!realloc_buffer(uart->rx_buffer, rxBufferSize)) {
+		if(!smg_uart_realloc_buffer(uart->rx_buffer, rxBufferSize)) {
 			delete uart;
 			return nullptr;
 		}
@@ -697,7 +622,7 @@ smg_uart_t* smg_uart_init_ex(const smg_uart_config_t& cfg)
 	}
 
 	if(smg_uart_tx_enabled(uart)) {
-		if(!realloc_buffer(uart->tx_buffer, txBufferSize)) {
+		if(!smg_uart_realloc_buffer(uart->tx_buffer, txBufferSize)) {
 			delete uart->rx_buffer;
 			delete uart;
 			return nullptr;
@@ -763,21 +688,6 @@ void smg_uart_uninit(smg_uart_t* uart)
 	delete uart->rx_buffer;
 	delete uart->tx_buffer;
 	delete uart;
-}
-
-smg_uart_t* smg_uart_init(uint8_t uart_nr, uint32_t baudrate, uint32_t config, smg_uart_mode_t mode, uint8_t tx_pin,
-						  size_t rx_size, size_t tx_size)
-{
-	smg_uart_config_t cfg = {.uart_nr = uart_nr,
-							 .tx_pin = tx_pin,
-							 .rx_pin = UART_PIN_DEFAULT,
-							 .mode = mode,
-							 .options = _BV(UART_OPT_TXWAIT),
-							 .baudrate = baudrate,
-							 .config = config,
-							 .rx_size = rx_size,
-							 .tx_size = tx_size};
-	return smg_uart_init_ex(cfg);
 }
 
 void smg_uart_set_config(smg_uart_t* uart, smg_uart_format_t config)
